Does not craete memory leak
